Moved tests within seviceProviderReferrals.spec.js and probationPractitionerReferrals.spec.js to individual files which are more granular to specific tasks connected with a referral.
    These were split into:

    actionPlan.spec.js
    appointmentFeedback.spec.js
    appointmentSchedulings.spec.js
    endOfServiceReport.spec.js
    referralDetails.spec.js
    referralProgress.spec.js
    reporting.spec.js
    reporting.sepc.js
    dashboard.spec.js

    The reason for splitting is:
    1. to make code easier to read and navigate.
    2. make writing cypress tests quicker; modifiying a test in a large cypress file will require running through the complete test to see if it works.
